+++ src/main/java/com/teatro/controlador/ShowControlador.java
@@ -44,7 +44,7 @@ public class ShowControlador {
 	public ResponseEntity<List<Show>> obtenerShows(@RequestParam("titulo") Optional<String> titulo,
 			@RequestParam("precio") Optional<Float> precio,
 			@RequestParam("fechaShow") Optional<LocalDateTime> fechaShow,
-			@RequestParam("categoriaId") Optional<Integer> categoriaId,
+			@RequestParam("categoriaId") Optional<Long> categoriaId,
 			@PageableDefault(size = 20, page = 0) Pageable pageable, HttpServletRequest request) {
 		Page<Show> shows = showServicio.buscarPorArgs(titulo, precio, fechaShow, categoriaId, pageable);
